+++ src/components/organisms/EntityDetails/EntityDetailsContent/Settings/SettingsTests/SettingsTests.tsx
@@ -1,6 +1,4 @@
 import {memo, useContext, useEffect, useMemo, useState} from 'react';
-import ReactFlow, {Controls, Edge, Node} from 'reactflow';
-import 'reactflow/dist/style.css';
 
 import {Form} from 'antd';
 
@@ -31,15 +29,10 @@ import {externalLinks} from '@utils/externalLinks';
 import {displayDefaultNotificationFlow} from '@utils/notification';
 import {convertTestSuiteV2ToV3, isTestSuiteV2} from '@utils/testSuites';
 
-import TestNode from './CustomNode';
-import DelayModal from './DelayModal';
-import IntersectionNode from './IntersectionNode';
-import {EmptyTestsContainer, ReactFlowContainer} from './SettingsTests.styled';
-import TestModal from './TestModal';
-
-interface ExtendedNode extends Node {
-  group?: number | string;
-}
+import DelayModal from './Modals/DelayModal';
+import TestModal from './Modals/TestModal';
+import {EmptyTestsContainer} from './SettingsTests.styled';
+import TestSuiteStepsFlow from './TestSuiteStepsFlow';
 
 const SettingsTests = () => {
   const {isClusterAvailable} = useContext(MainContext);
@@ -97,8 +90,6 @@ const SettingsTests = () => {
   const [currentGroup, setCurrentGroup] = useState<number | string>(0);
 
   const [steps, setSteps] = useState(initialSteps);
-  const [nodes, setNodes] = useState<ExtendedNode[]>([]);
-  const [edges, setEdges] = useState<Edge[]>([]);
 
   const wasTouched = steps !== initialSteps;
 
@@ -115,75 +106,6 @@ const SettingsTests = () => {
 
     setCurrentGroup(group);
   };
-
-  const removeNode = (id: string, group: number) => {
-    if (steps[group].length === 1) {
-      setSteps([...steps.slice(0, group), ...steps.slice(group + 1)]);
-    } else {
-      setSteps([...steps.slice(0, group), steps[group].filter(item => item.id !== id), ...steps.slice(group + 1)]);
-    }
-  };
-
-  // add nodes
-  useEffect(() => {
-    const newNodes: ExtendedNode[] = [];
-
-    steps.forEach((step, index) => {
-      step.forEach((item: LocalStep, itemIndex: number) =>
-        newNodes.push({
-          type: 'test',
-          id: item.id!,
-          data: {...item, removeNode, group: index},
-          position: {x: index * 450, y: itemIndex * 100},
-          group: index,
-        })
-      );
-
-      newNodes.push({
-        type: 'intersection',
-        id: nanoid(),
-        position: {x: index * 450 + 135, y: step.length * 100 - 20},
-        data: {showModal, group: index, withoutHandles: true},
-      });
-
-      const intersectionGroup = `${index}-intersection`;
-      newNodes.push({
-        type: 'intersection',
-        id: nanoid(),
-        position: {x: (index + 1) * 450 - 90, y: 11},
-        data: {showModal, group: intersectionGroup},
-        group: intersectionGroup,
-      });
-    });
-
-    setNodes(newNodes);
-  }, [steps]);
-
-  // add edges
-  useEffect(() => {
-    const newEdges: Edge[] = [];
-
-    nodes.forEach(node => {
-      const currentId = node.id;
-
-      const nextIntersection =
-        typeof node.group === 'number' && nodes.find(x => x.group === `${node.group}-intersection`);
-
-      if (nextIntersection) {
-        newEdges.push({id: `${currentId}-${nextIntersection.id}`, source: currentId, target: nextIntersection.id});
-      } else {
-        const filteredNodes = nodes.filter(x => x.group === Number(String(node.group).split('-')[0]) + 1);
-
-        filteredNodes.forEach(nextNode => {
-          newEdges.push({id: `${currentId}-${nextNode.id}`, source: currentId, target: nextNode.id});
-        });
-      }
-    });
-
-    setEdges(newEdges);
-  }, [nodes]);
-
-  const nodeTypes = useMemo(() => ({test: TestNode, intersection: IntersectionNode}), []);
 
   const addNode = (data: number | {name: string; type: string}) => {
     const item =
@@ -214,7 +136,7 @@ const SettingsTests = () => {
             test: item.test,
           };
         }),
-        stopOnFailure: entityDetails?.steps ? entityDetails?.steps[i]?.stopOnFailure || false : false,
+        stopOnFailure: Boolean(entityDetails?.steps?.[i]?.stopOnFailure),
       };
     });
 
@@ -258,11 +180,7 @@ const SettingsTests = () => {
             </Button>
           </EmptyTestsContainer>
         ) : (
-          <ReactFlowContainer>
-            <ReactFlow nodeTypes={nodeTypes} nodes={nodes} edges={edges}>
-              <Controls position="bottom-right" showInteractive={false} />
-            </ReactFlow>
-          </ReactFlowContainer>
+          <TestSuiteStepsFlow steps={steps} setSteps={setSteps} showModal={showModal} />
         )}
         <DelayModal
           isDelayModalVisible={isDelayModalVisible}
